src/api/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
anyjson==0.3.3
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
boto3==1.26.61
botocore==1.29.61
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
chroma-hnswlib==0.7.2
chromadb==0.4.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
coloredlogs==15.0.1
config==0.5.1
configparser==5.3.0
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
Django==4.1.5
django-anymail==9.2
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-otp==1.1.6
django-phonenumber-field==7.0.2
django-softdelete==0.10.5
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
django-ses==3.4.1
django-timezone-field==5.0
djangorestframework==3.14.0
docopt==0.6.2
dynaconf==3.1.11
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.11.0
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
huggingface-hub==0.13.4
humanfriendly==10.0
hvac==1.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
jax==0.4.13
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras
kiwisolver==1.4.4
knox==0.1.14
kombu==5.2.4
langchain==0.0.249
langsmith==0.0.16
libclang==16.0.0
lit==16.0.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.3.1
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.0.10
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
phonenumbers==8.13.4
Pillow==9.5.0
posthog==3.0.1
prompt-toolkit==3.0.36
protobuf==4.23.3
psutil==5.9.4
psycopg2==2.9.5
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.2.3
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.12
pyhcl==0.4.4
pypandoc==1.11
pyparsing==3.0.9
PyPika==0.48.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
shortuuid==1.0.11
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.3
starlette==0.27.0
stripe==5.3.0
sumy==0.11.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2022.7
unstructured==0.8.8
urllib3==1.26.14
uvicorn==0.23.2
uvloop==0.17.0
validators==0.20.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
wrapt
xlrd==2.0.1
XlsxWriter==3.0.8
yarl==1.9.2
zipp==3.15.0
fuzzywuzzy==0.18.0
spacy
pdfplumber
fitz
PyMuPDF
flair
exponent_server_sdk
azure-communication-email